LPS-168592 Move also test cases
