+++ modules/apps/layout/layout-impl/build.gradle
@@ -54,4 +54,6 @@ dependencies {
 	compileOnly project(":core:petra:petra-function")
 	compileOnly project(":core:petra:petra-reflect")
 	compileOnly project(":core:petra:petra-string")
+
+	testCompile group: "com.github.everit-org.json-schema", name: "org.everit.json.schema", version: "1.11.1"
 }
