+++ modules/apps/layout/layout-impl/src/test/resources/com/liferay/layout/internal/importer/validator/dependencies/display_page_template_invalid_extra_properties.json
