+++ modules/apps/layout/layout-impl/src/test/resources/com/liferay/layout/internal/importer/validator/dependencies/page_definition_invalid_drop_zone_allowed_fragments_unallowed_fragments.json
