+++ modules/apps/layout/layout-impl/src/test/resources/com/liferay/layout/internal/importer/validator/dependencies/page_definition_invalid_root_extra_properties.json
