+++ modules/apps/layout/layout-impl/src/test/resources/com/liferay/layout/internal/importer/validator/dependencies/page_definition_valid_fragment_field_text_fragment_link_href_fragment_mapped_value_mapping_item_reference_class_field_reference.json
