+++ modules/apps/layout/layout-impl/src/test/resources/com/liferay/layout/internal/importer/validator/dependencies/page_template_collection_invalid_extra_properties.json
